requirements.txt
beautifulsoup4==4.13.3
matplotlib==3.7.1
mlflow==2.13.2
numpy==1.24.2
pandas==2.0.0
Pillow==9.5.0
Pillow==11.1.0
PyMySQL==1.1.1
python_dateutil==2.8.2
Requests==2.32.3
requests_ntlm==1.3.0
ruptures==1.1.9
scikit_learn==0.24.1
scipy==1.15.2
statsmodels==0.14.4
tqdm==4.65.0
trino==0.326.0
